Update gasfvf.py
There was an error with the unit for gasfvf function, unit should be reservoir cubic feet per standard cubic feet, not reservoir barrels.
Also, I merged both functions into one as I think its better for interfacing with users of the function. With respect to interfacing with users, I think it will be better to use strings that reservoir engineers are familiar with as the unit parameters.
Lastly, I added a form of validation to check for errors in unit input.